+++ src/game/xui/AirportGui.java
@@ -185,7 +185,7 @@ public class AirportGui
 	    // strings such as 'Size' and 'Coverage Area'
 			Gfx.DrawStringCentered(74, 16, Str.STR_305B_SIZE, 0);
 			Gfx.DrawStringCentered(74, 78, Str.STR_3066_COVERAGE_AREA_HIGHLIGHT, 0);
-			MiscGui.DrawStationCoverageAreaText(2, 104, (int)-1, rad);
+			MiscGui.DrawStationCoverageAreaText(2, 104, -1, rad);
 			break;
 		}
