+++ src/game/xui/RoadVehGui.java
@@ -580,7 +580,7 @@ public class RoadVehGui
 				DrawRoadVehImage(v, x+24, y, w.as_traindepot_d().sel);
 
 				Global.SetDParam(0, v.getUnitnumber().id);
-				Gfx.DrawString(x, y+2, (int)(v.getMax_age()-366) >= v.getAge() ? Str.STR_00E2 : Str.STR_00E3, 0);
+				Gfx.DrawString(x, y+2, (v.getMax_age()-366) >= v.getAge() ? Str.STR_00E2 : Str.STR_00E3, 0);
 
 				Gfx.DrawSprite( v.isStopped() ? Sprite.SPR_FLAG_VEH_STOPPED : Sprite.SPR_FLAG_VEH_RUNNING, x + 16, y);
 
@@ -664,8 +664,8 @@ public class RoadVehGui
 
 	/**
 	 * Clones a road vehicle
-	 * @param *v is the original vehicle to clone
-	 * @param *w is the window of the depot where the clone is build
+	 * @param v is the original vehicle to clone
+	 * @param w is the window of the depot where the clone is build
 	 */
 	static void HandleCloneVehClick(final Vehicle  v, final Window  w)
 	{
